Do not do flushing seeks on set rate.
In order to set a rate a seek must be performed.
The best option is to do the flushing seek as it can be done
both in PLAYING and PAUSED state and works most reliably.
The problem is that these obviously flush all the data and
we have no way of asking for the same data again.
Non flushing seek can be done only in PLAYING but even then
it deadlocks/freezes on some platforms. There's no other way.
This is why custom "rate" property is required to be supported
by sinks or as a last resort platform is asked to do the task
instead of performing seeks.